+++ frontend/src/routes.js
@@ -3,7 +3,7 @@ import Icon from "@mui/material/Icon";
 
 // Pages
 import SignIn from "layouts/pages/authentication/sign-in";
-import PageHeaders from "./layouts/sections/page-sections/page-headers";
+import PageHeaders from "./layouts/sections/page-sections/recursos";
 
 const routes = [
   {
@@ -12,17 +12,17 @@ const routes = [
     collapse: [
       {
         name: "Recursos informatius",
-        route: "/sections/page-sections/page-headers",
+        route: "/sections/page-sections/recursos",
         component: <PageHeaders />,
       },
       {
         name: "Recursos visuals",
-        route: "/sections/page-sections/page-headers",
+        route: "/sections/page-sections/recursos",
         component: <SignIn />,
       },
       {
         name: "Recursos audiovisuals",
-        route: "/sections/page-sections/page-headers",
+        route: "/sections/page-sections/recursos",
         component: <SignIn />,
       },
     ],